backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
google-auth==2.3.3
idna==3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
requests==2.26.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
